% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracer.R
\name{Check existence of a trace}
\alias{Check existence of a trace}
\alias{has_trace}
\title{Check if the trace would exist after the call}
\usage{
has_trace(tracer, fun, pkg = NULL, args = list(), globals = list())
}
\description{
Check if the trace would exist after the call
}
