% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genthat.R
\name{compare_test_results}
\alias{compare_test_results}
\title{Returns comparison of two results returned by \code{run_generated_tests()}.}
\usage{
compare_test_results(resA, resB)
}
\arguments{
\item{resA}{first result}

\item{resB}{second result}
}
\value{
comparison structure
}
\description{
Returns comparison of two results returned by \code{run_generated_tests()}.
}

