% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{concat}
\alias{concat}
\title{Expects string arguments and returns the concatenation of all of them.}
\usage{
concat(..., sep = "")
}
\arguments{
\item{sep}{optional separator}
}
\value{
character scalar
}
\description{
Every argument must me a single string.
}

