% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{concatVec}
\alias{concatVec}
\title{Expects exactly one character vector argument and returns all the elements concatenated together.}
\usage{
concatVec(xs, sep = "")
}
\arguments{
\item{sep}{optional separator}
}
\value{
character scalar
}
\description{
Every argument must me a single string.
}

