% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracer.R
\name{create_tracer}
\alias{create_tracer}
\title{Creates a new tracer of the given type.}
\usage{
create_tracer(type = getOption("genthat.tracer_type"), ...)
}
\arguments{
\item{type}{the type of the tracer, can e either \code{set} of
\code{sequence}. The default is what is set in
\code{genthat.tracer_type} option.}

\item{...}{additional parameters to be passed to the function creating the
#actual tracer}
}
\description{
Creates either a set tracer that only stores unique traces (unique
combination of function/package name, argument values and return value) or
sequence tracer that stores all traces.
}
\seealso{
\code{\link{create_set_tracer}}

\code{\link{create_sequence_tracer}}
}
