% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorate.R
\name{decorate_environment}
\alias{decorate_environment}
\title{Decorates functions in an environment}
\usage{
decorate_environment(envir)
}
\arguments{
\item{envir}{an environment that shall be decorated}
}
\description{
Decorates all symbols form the given environment for which `is.function` is `TRUE`.
}
