% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture.R
\name{decorate_exported}
\alias{decorate_exported}
\title{Decorates package's exported functions.}
\usage{
decorate_exported(package, functions = NULL, ..., all = FALSE)
}
\arguments{
\item{package}{name of package to look for function}

\item{functions}{character vector of all the functions to decorate}

\item{all}{decorate all the exported functions}
}
