% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture.R
\name{decorate_function_env}
\alias{decorate_function_env}
\title{Decorates function bound in environment}
\usage{
decorate_function_env(func_names, env = sys.frame(-1))
}
\arguments{
\item{env}{environment containging binding}

\item{func_name}{key of binding}
}
