% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture.R
\name{decorate_function_val}
\alias{decorate_function_val}
\title{Decorates function to capture calls and return values}
\usage{
decorate_function_val(func, func_name, tracer.expr)
}
\arguments{
\item{func}{function value}

\item{func_name}{function name}
}
\description{
More sane replacement of the base::trace function.
}

