% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture.R
\name{decorate_function_val__}
\alias{decorate_function_val__}
\title{Returns traced version of original function.}
\usage{
decorate_function_val__(func, func_label, enter_function, exit_function)
}
\arguments{
\item{func}{function value}

\item{func_label}{name that will be recorded in traces}

\item{enter_function}{name that will be recorded in traces}

\item{exit_function}{name that will be recorded in traces}
}
\description{
More sane replacement of the base::trace function.
}
