% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorate.R
\name{decorate_functions}
\alias{decorate_functions}
\title{Decorates given functions}
\usage{
decorate_functions(...)
}
\arguments{
\item{...}{the functions that shall be decorated}
}
\description{
Given functions will be decorated in their defining environment.
}
