% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture.R
\name{decorate_hidden_functions}
\alias{decorate_hidden_functions}
\title{Decorates package's exported functions.}
\usage{
decorate_hidden_functions(package)
}
\arguments{
\item{package}{name of package}

\item{functions}{character vector of all the functions to decorate}

\item{all}{decorate all the exported functions}
}
