% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-package.R
\name{download_package}
\alias{download_package}
\title{Downloads R package}
\usage{
download_package(package, destdir, version = NULL,
  repos = getOption("repos"), type = "source", extract = TRUE,
  force = FALSE, quiet = TRUE, ...)
}
\description{
Downloads R package.
}
