% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture.R
\name{enter_function}
\alias{enter_function}
\title{Write down capture information}
\usage{
enter_function(fname, args.env, call_id)
}
\arguments{
\item{fname}{function name}

\item{args.env}{environment to read arguments to function call from}
}
\description{
This function is respinsible for writing down capture information for decorated function calls.
}
\seealso{
Decorate
}

