% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-package-code.R
\name{extract_package_code}
\alias{extract_package_code}
\title{Extracts code from examples, tests and/or vignettes for a given package.}
\usage{
extract_package_code(package, dir, lib_paths = NULL, types = c("examples",
  "tests", "vignettes", "all"), output_dir = ".", filter = NULL)
}
\arguments{
\item{package}{the name of the package.}

\item{dir}{the path to the directory containing the source of the package}

\item{lib_paths}{a character vector describing the location of R library
trees to search through when locating `package`, or `NULL`. The default
value of `NULL`` corresponds to checking the loaded namespace, then all
libraries currently known in `.libPaths()`.}

\item{types}{which code artifacts to extract, `tests`, `vignettes`,
`examples`, `all`.}

\item{output_dir}{target directory where to put the extracted source. Default
is the current directory.}

\item{filter}{a regular expression matching desired filenames without the
'.R' extension.}
}
\value{
Returns a list where each element is a character vector of file names
    where the code has been extracted. The element names correspond to the
    elements in the types parameter.
}
\description{
This function extracts code from package examples, tests and/or vignettes as
runnable R scripts. It can extract the code for already installed packages,
given that it has been installed with \code{INSTALL_opts=c("--examples",
"--install-tests")} and with built vignettes. Alternatively, it can extract
the same from a package source. In such a case it will attempt to temporarily
install the package.
}
