% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{format_value}
\alias{format_value}
\title{Format a given value into R code string}
\usage{
format_value(x, ...)
}
\arguments{
\item{x}{the value to format
/}
}
\description{
Creates an R code snippet such that `isTRUE(all.equals(eval(format_value(x)), x))`.
}
