% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnecessary.R
\name{gen_from_cran}
\alias{gen_from_cran}
\title{Generate tests for given package}
\usage{
gen_from_cran(package, version, verbose = TRUE, source.dir = "sources",
  out.dir = "generated_tests", filter = FALSE, clear_capture = FALSE,
  include.manPages = TRUE, include.vignettes = TRUE, include.tests = TRUE)
}
\arguments{
\item{package}{name of package}
}
\description{
Generates tests from package downloaded from CRAN
}

