% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genthat.R
\name{gen_from_package}
\alias{gen_from_package}
\title{Extract unit tests for a package}
\usage{
gen_from_package(path, from = NULL, types = c("examples", "tests",
  "vignettes", "all"), filter = NULL, ...)
}
\arguments{
\item{path}{file path to the package.}

\item{from}{a character vector of package names, from which code shall be
used to extract the code from. Default \code{NULL} stands for the package
in `path`. For non-\code{NULL} values, the packages shall be installed
with the appropriate options, i.e. \code{INSTALL_opts('--example',
'--install-tests')} in order to contain the required code. For
\code{NULL}, the package will be installed with the appropriate options
based on the \code{types}.}

\item{types}{which code artifacts to run from the `from` packages,
'examples', 'tests', 'vignettes' or 'all'}
}
\description{
This function extracts units tests for a given package on the `path`. By
default it runs package examples, tests and vignettes.
}
\examples{
\dontrun{gen_from_package(".")}

# in case the package has been already installed with necessary artifacts
\dontrun{gen_from_package(find.package("my-packages"))}

}
