% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_tc}
\alias{generate_tc}
\title{Generate test case from trace}
\usage{
generate_tc(trace)
}
\arguments{
\item{trace}{trace value}
}
\description{
This function generates a test case from the passed trace.
}
