% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_test}
\alias{generate_test}
\title{generate test from a trace}
\usage{
generate_test(trace, ...)
}
\value{
a data frame or a tibble with the following
fun        : chr
pkg        : chr
code       : chr (can be NA in which case gen_error must be a chr)
error      : chr (can be NA in which case code must be NA)
elapsed    : numeric (can be NA)
}
\description{
given a trace, it generates a test
}
