% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_test_code}
\alias{generate_test_code}
\title{Generate test case code from a trace}
\usage{
generate_test_code(trace)
}
\arguments{
\item{trace}{trace value}
}
\description{
Given a genthat trace it generates a corresponding test case
}
