% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_test_file}
\alias{generate_test_file}
\title{generate test from a trace}
\usage{
generate_test_file(trace, output_dir, ...)
}
\value{
a data frame with the following
pkg        : chr
fun        : chr
filename   : chr (can be NA in which case error must be a chr)
error      : chr (can be NA in which case code must be NA)
elapsed    : numeric (can be NA)
}
\description{
given a trace, it generates a test and stores it in a file
}
