% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_tests}
\alias{generate_tests}
\title{Generates test cases from traces}
\usage{
generate_tests(traces, include_trace_dump = FALSE)
}
\arguments{
\item{traces}{from which to generate test cases}
}
\description{
Generates tests cases from the captured traces.
}
