% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture.R
\name{genthat_trace}
\alias{genthat_trace}
\title{Decorates function to capture calls and return values}
\usage{
genthat_trace(func, tracer)
}
\arguments{
\item{func}{function name as a character string}

\item{tracer}{expression to insert at beginning of function body}
}
\description{
More sane replacement of the base::trace function.
}

