% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorate.R
\name{get_decorations}
\alias{get_decorations}
\title{Returns decorated functions}
\usage{
get_decorations()
}
\value{
a list where names are fully-qualified function names and values are
    the decorated functions
}
\description{
Return a list of the decorated functions.
}
