% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_default_arguments}
\alias{get_default_arguments}
\title{Gets the arguments of the function that use their default argument

Such an argument is part of the formal parameter list, with a default alue, but is not
specified in the call.

It means it has a value in the local environment but is considered to be missing.}
\usage{
get_default_arguments(call, formal_args)
}
\description{
Gets the arguments of the function that use their default argument

Such an argument is part of the formal parameter list, with a default alue, but is not
specified in the call.

It means it has a value in the local environment but is considered to be missing.
}
