% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-sources-archive.R
\name{get_package_from_archive}
\alias{get_package_from_archive}
\title{Extracts package from an archive and installs all the dependencies.}
\usage{
get_package_from_archive(path, dest.dir = "sources", quiet = TRUE)
}
\arguments{
\item{path}{Path to \emph{.tar.gz} file.}

\item{dest.dir}{Destination folder}

\item{quiet}{Suppress messages.}
}
\value{
Returns a value of type \code{\link{fetchResult}}:
}
\description{
Extracts package from an archive and installs all the dependencies.
}
\examples{
get_package_from_archive("./vcrpart_0.3-3.tar.gz")
get_package_from_archive("./vcrpart_0.3-3.tar.gz", dest.dir="./package_sources", quiet=FALSE)
}

