% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-sources-cran.R
\name{get_package_from_cran}
\alias{get_package_from_cran}
\title{Fetches a package from CRAN and installs all the dependencies.}
\usage{
get_package_from_cran(pkg_name, version, archive.dir = tempdir(),
  dest.dir = "sources", quiet = TRUE,
  CRAN_repo = "http://cran.us.r-project.org")
}
\arguments{
\item{pkg_name}{Name of package in CRAN repo.}

\item{version}{Optional version name. If not specified, the most recent version is fetched.}

\item{archive.dir}{Where to download the archive files. If not specified a temporary folder is used and it is deleted upon completion.}

\item{dest.dir}{Destination folder}

\item{quiet}{Suppress messages.}
}
\value{
Returns a value of type \code{\link{fetchResult}}:
}
\description{
Fetches a package from CRAN and installs all the dependencies.
}
\examples{
get_package_from_cran("vcrpart")
get_package_from_archive("vcrpart", "0.3-3", dest.dir="./package_sources")
}

