% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-sources-result.R
\name{get_package_result}
\alias{get_package_result}
\title{Result of a \code{get_package_...(...)} function.}
\usage{
get_package_result(pkgName, packageDir)
}
\arguments{
\item{pkgName}{Package name.}

\item{packageDir}{Location of package sources.}
}
\value{
Returns a \code{get_package_result} instance.
}
\description{
Result of a \code{get_package_...(...)} function.
}
\examples{
get_package_result("vcrpart", "./sources/vcrpart")
}

