% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-package.R
\name{get_package_version}
\alias{get_package_version}
\title{Gets the latest package version}
\usage{
get_package_version(package, repos = getOption("repos"), type = "source")
}
\description{
Returns the latest version of a given package that is available in the given repository.
}
