% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracer.R
\name{get_tracer}
\alias{get_tracer}
\title{Gets the currently used tracer}
\usage{
get_tracer()
}
\value{
the current tracer
}
\description{
Returns the current tracer. If no such tracer exists, it will create one by
calling \code{create_tracer()} with no arguments, i.e. using the default
settings.
}
\seealso{
\code{\link{create_tracer()}}
}
