% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorate.R
\name{is_decorated}
\alias{is_decorated}
\title{Checks whether a function has been already decorated}
\usage{
is_decorated(fun)
}
\arguments{
\item{fun}{function value}
}
\description{
Checks whether the given function has been decorated by `decorate_functions` call.
}
