% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{isAccessibleFunction}
\alias{isAccessibleFunction}
\title{Do we have access to the function in test generation phase, to recompute the return value from the arguments?}
\usage{
isAccessibleFunction(function.name)
}
\arguments{
\item{function.name}{string name of binding}
}
\value{
boolean value signaling whether evaluating the supplied function name in the global scope yields the function or not.
}
\description{
Do we have access to the function in test generation phase, to recompute the return value from the arguments?
}
\examples{
isAccessibleFunction("myFn1") == FALSE
isAccessibleFunction("ggplot2::ggplot") == TRUE
}

