% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{link_environments}
\alias{link_environments}
\title{Links the environments of the surrounding functions}
\usage{
link_environments(env = parent.frame(), parent = parent.env(env),
  .fun_filter = is.local_closure)
}
\arguments{
\item{env}{the environment in which to look for functions}

\item{parent}{the environment to use as the parent environment of the functions}
}
\description{
Sets the parent environment of all the functions defined in the given environment `env` to `parent`.
}
