% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-sources-cran.R
\name{list_cran_package_versions}
\alias{list_cran_package_versions}
\title{Lists the available versions of a package on the given CRAN mirror.}
\usage{
list_cran_package_versions(package,
  CRAN_repo = "http://cran.us.r-project.org")
}
\arguments{
\item{package}{Package name.}

\item{CRAN_repo}{CRAN mirror url.}
}
\value{
Returns a list of version names.
}
\description{
Lists the available versions of a package on the given CRAN mirror.
}
\examples{
list_cran_package_versions("ggplot2")
}

