% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated-tests.R
\name{parseGeneratedTest}
\alias{parseGeneratedTest}
\title{Accepts one string representing a call to \code{test_that()} and returns an appropriate data structure.}
\usage{
parseGeneratedTest(text)
}
\arguments{
\item{text}{String representation of test.}
}
\description{
Accepts one string representing a call to \code{test_that()} and returns an appropriate data structure.
}
\examples{
parseGeneratedTest("\\n\\ttest_that('test-42', { expected <- 42; expect_equal(the_truth_function("anything"), expected) })\\n\\n")
}

