% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-result.R
\name{print.gen_result}
\alias{print.gen_result}
\title{Generate tests for given package}
\usage{
\method{print}{gen_result}(x)
}
\arguments{
\item{package}{name of package}
}
\description{
Generates tests from package downloaded from CRAN
}

