% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genthat.R
\name{process_traces}
\alias{process_traces}
\title{Process traces based on given action}
\usage{
process_traces(traces, output_dir, action)
}
\arguments{
\item{file}{that has been running}
}
\description{
Process traces based on given action
}
