% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorate.R
\name{reset_function}
\alias{reset_function}
\title{Resets decorated function back to its original}
\usage{
reset_function(fun, name = substitute(fun), decorator = get_decorator(),
  env = parent.frame())
}
\description{
Reverts decorated function back to their state they were before calling `decorate_function`.
}
