% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorate.R
\name{reset_functions}
\alias{reset_functions}
\title{Resets decorated function back to its original}
\usage{
reset_functions(..., in_env = parent.frame())
}
\description{
Reverts decorated functions back to their state they were before calling `decorate_functions`.
}
