% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-generated-tests.R
\name{run_generated_tests}
\alias{run_generated_tests}
\title{status code:
 0: OK
-1: Test could not be run - i.e. a syntax error
 1: Test failed
 2: Test threw an exception
 3: Test did not contain any tests (nb field in the testthat_result was 0)}
\usage{
run_generated_tests(tests, quiet = TRUE,
  show_progress = isTRUE(getOption("genthat.show_progress")))
}
\description{
status code:
 0: OK
-1: Test could not be run - i.e. a syntax error
 1: Test failed
 2: Test threw an exception
 3: Test did not contain any tests (nb field in the testthat_result was 0)
}
