% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-package.R
\name{run_package}
\alias{run_package}
\title{run package code}
\usage{
run_package(pkg, types = c("examples", "tests", "vignettes"), fork = TRUE,
  quiet = TRUE, clean = TRUE, lib_path = .libPaths(), ...)
}
\arguments{
\item{...}{commentDontrun, commentDonttest}
}
\description{
run package examples, tests and/or vignettes
}
