% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{save_test}
\alias{save_test}
\title{Saves 'code' for the give 'pkg'::'fun' into 'output_dir'}
\usage{
save_test(pkg, fun, code, output_dir)
}
\arguments{
\item{tests}{this should be a data.frame with class genthat_tests, a result
from calling `generate_tests`.}
}
\description{
The test and its `ext`` counterpart will be stored in
`getwd()/<pkg>/<fun>/test-<n>.{R,ext}`. Where `<n>` is the next increment of
the number of `test-<i>.R` files in the directory.
}
