% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracer.R
\name{set_tracer}
\alias{set_tracer}
\title{Makes the given \code{tracer} default}
\usage{
set_tracer(tracer)
}
\arguments{
\item{tracer}{the tracer that shall become the default one}
}
\value{
the previous tracer or \code{NULL} if no tracer has been set
    yet.
}
\description{
Sets the given tracer to be the one used for tracing from now on.
}
