% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture.R
\name{stop_capture_all}
\alias{stop_capture_all}
\title{Stops capturing all currently captured functions.}
\usage{
stop_capture_all(verbose = FALSE)
}
\arguments{
\item{verbose}{TRUE to display additional information}
}
\description{
Remove tracing functionality for all the functions
}

