% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracer.R
\name{store_trace}
\alias{store_trace}
\title{Stores trace}
\usage{
store_trace(tracer, trace)
}
\arguments{
\item{tracer}{the tracer to store the \code{trace} into}

\item{trace}{the trace to be store}
}
\description{
Stores the given \code{trace} to the given \code{tracer}.
}
