% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{strip_prefix}
\alias{strip_prefix}
\title{Removes prefixes and quote from line}
\usage{
strip_prefix(prefix, line)
}
\arguments{
\item{prefix}{prefix}

\item{line}{input line}
}
\description{
Used for processing capture file information. Deletes prefixes to get essential information
}
\seealso{
ProcessClosure
}

