% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture.R
\name{undecorate_all}
\alias{undecorate_all}
\title{Undecorate all functions.}
\usage{
undecorate_all()
}
\description{
Undecorate all functions.
Functions decorated through `decorate_function_val()` are not undecorated.
}
