% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorate.R
\name{decorate_environment}
\alias{decorate_environment}
\title{Decorates functions in an environment}
\usage{
decorate_environment(env, decorator = get_decorator(),
  record_fun = substitute(genthat:::record_trace), exclude = character())
}
\arguments{
\item{envir}{an environment that shall be decorated or a character scalar
that will be resolved using `as.environment("package:....")`. In the
latter case, the package will also be loaded in the case it has not been
loaded yet.}
}
\description{
Decorates all symbols form the given environment for which
    `is.function` is `TRUE`.
}
