% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorate.R
\name{decorate_function}
\alias{decorate_function}
\title{Decorates given functions}
\usage{
decorate_function(fun, name = substitute(fun),
  record_fun = substitute(genthat:::record_trace),
  decorator = get_decorator(), env = parent.frame())
}
\arguments{
\item{fun}{the function that shall be decorated}
}
\description{
Given function will be decorated in their defining environment.
}
