% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture.R
\name{decorate_function}
\alias{decorate_function}
\title{Decorates function bound in environment}
\usage{
decorate_function(func_names, env = environment())
}
\arguments{
\item{env}{environment containging binding}

\item{func_name}{key of binding}
}
