% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorate.R
\name{decorate_function}
\alias{decorate_function}
\title{Decorates given function}
\usage{
decorate_function(fun, name = substitute(fun))
}
\arguments{
\item{fun}{the function that shall be decorated}

\item{name}{name of the function}
}
\description{
Given function will be decorated in its defining environment.
}
