% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorate.R
\name{decorate_functions}
\alias{decorate_functions}
\title{Decorates given functions}
\usage{
decorate_functions(..., in_env = parent.frame(),
  .recorder = substitute(genthat:::record_trace))
}
\arguments{
\item{...}{the functions that shall be decorated

decorate_functions(list(ls, ...))
decorate_functions(list("ls", ...))
decorate_functions(c("ls", ...))
decorate_functions(ls, ...)
decorate_functions("ls", ...)}
}
\description{
Given functions will be decorated in their defining environment.
}
