% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture.R
\name{decorate_functions}
\alias{decorate_functions}
\title{Decorate functions}
\usage{
decorate_functions(...)
}
\description{
API function to decorate functions.
Usages
------
decorate_functions(fn1)
decorate_functions(fn1, fn2)
decorate_functions("fn1", env=environment())
decorate_functions("fn1", "fn2", env=environment())
decorate_functions("fn1", "fn2", package="package_name")
decorate_functions(package="package_name") # all from this package
decorate_functions(env=some_environment) # all from some_environment
decorate_functions(package="package_name", include_hidden=TRUE) # all including hidden
decorate_functions("hidden_fn1", "hidden_fn2", package="package_name", include_hidden=TRUE)
}
