% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genthat.R
\name{gen_from_package}
\alias{gen_from_package}
\title{Generate test cases for a package}
\usage{
gen_from_package(package, output_dir = "generated_tests", types = c("tests",
  "vignettes", "examples"), clean = FALSE, quiet = TRUE)
}
\description{
Decorates all functions in a package and then generates test cases based on
the code contained in the package examples, vignettes and tests.
}
