% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{gen_tests}
\alias{gen_tests}
\title{Generates tests from captured information.}
\usage{
gen_tests(output_dir = "generated_tests")
}
\arguments{
\item{output_dir}{Directory to which the tests should be generated.}
}
\description{
This function takes the tracing information collected by capture and generates
testthat compatible testcases.
}
