% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_tc}
\alias{generate_tc}
\title{Generates a testcase for closure function}
\usage{
generate_tc(func, argv, retv)
}
\arguments{
\item{func}{function name}

\item{argv}{input arguments for a closure function call}
}
\description{
This function generates a test case for builtin function using supplied arguments. All elements should be given as text.
}
\seealso{
test_gen ProcessClosure
}

