% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_test}
\alias{generate_test}
\title{generate test from a trace}
\usage{
generate_test(trace, ...)
}
\value{
a data frame or a tibble with the following
trace      : chr
fun        : chr
code       : chr (can be NA)
error      : chr (can be NA)
elapsed    : numeric
}
\description{
given a trace, it generates a test
}
