% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_tests}
\alias{generate_tests}
\title{Generates test cases from traces}
\usage{
generate_tests(traces, quiet = TRUE, ...)
}
\arguments{
\item{traces}{from which to generate test cases}

\item{...}{additional arguments supplied to `generate_test_code` function.}
}
\value{
a data frame or a tibble with the following
fun       : chr
pkg       : chr
code      : chr (can be NA in which case gen_error must be a chr)
gen_error : chr (can be NA in which case code must be a NA)
elapsed   : dbl
}
\description{
Generates tests cases from the captured traces.
}
