% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_tests}
\alias{generate_tests}
\title{Generates test cases from traces}
\usage{
generate_tests(traces, output_dir = "generated_tests")
}
\arguments{
\item{traces}{from which to generate test cases}

\item{output_dir}{target directory where to place generated test}
}
\description{
Generates tests cases from the captured traces.
}
