% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_tests}
\alias{generate_tests}
\title{Generates test cases from traces}
\usage{
generate_tests(traces,
  show_progress = isTRUE(getOption("genthat.show_progress")), ...)
}
\arguments{
\item{traces}{from which to generate test cases}

\item{show_progress}{(log) show progress during test generation}

\item{...}{additional arguments supplied to `generate_test_code` function.}
}
\value{
a data frame or a tibble with the following
fun        : chr
trace      : chr
trace_type : chr
code       : chr (can be NA)
error      : chr (can be NA)
}
\description{
Generates tests cases from the captured traces.
}
