% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorate.R
\name{reset_function}
\alias{reset_function}
\title{Resets decorated function back to its original}
\usage{
reset_function(fun, env = parent.frame())
}
\arguments{
\item{fun}{function or function name}
}
\description{
Reverts decorated function back to their state they were before calling
\code{decorate_function}.
}
