% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genthat.R
\name{run_generated_tests}
\alias{run_generated_tests}
\title{Run tests generated by genthat.}
\usage{
run_generated_tests(dir)
}
\arguments{
\item{Path}{to directory containing the generated tests.}
}
\description{
Run tests generated by genthat.
}

