% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_package}
\alias{run_package}
\title{@param ... commentDontrun, commentDonttest}
\usage{
run_package(pkg, types = c("examples", "tests", "vignettes"), clean = TRUE,
  lib_path = .libPaths(), ...)
}
\description{
@param ... commentDontrun, commentDonttest
}
