% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-package.R
\name{run_package}
\alias{run_package}
\title{run package code}
\usage{
run_package(pkg, pkg_dir = find.package(pkg), types = c("examples", "tests",
  "vignettes", "all"), working_dir = tempfile(pattern = "run_package"),
  quiet = TRUE, runner = run_r_script)
}
\description{
run package examples, tests and/or vignettes
}
