% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genthat.R
\name{trace_package}
\alias{trace_package}
\title{Decorate all functions from given package and runs package code}
\usage{
trace_package(pkg, types = c("examples", "tests", "vignettes", "all"),
  output_dir = ".", filter = NULL, working_dir = tempfile(pattern =
  "genthat-trace-"), decorator = "onexit", tracer = "set", quiet = TRUE,
  lib_paths = NULL)
}
\arguments{
\item{output_dir}{the name of the directory where to output traces or NULL if traces should not be saved}
}
\description{
Decorates all functions in a package and then generates test cases based on
the code contained in the package examples, vignettes and tests.
}
